change sink to integer (to fix switch cases), usage instructions change url, don't start debugging on enqueue

+++ Inkybot.Dofus/Contracts/DofusMagingAI.cs
@@ -29,7 +29,7 @@ namespace Inkybot.Dofus.Contracts
         
         private MageConfigManager MageConfigManager = null!;
 
-        protected decimal Sink => SinkProvider.Sink;
+        protected int Sink => SinkProvider.Sink;
         
         /**
          * <summary>
